Add Standalone Compat class
